+++ packages/terra-worklist-data-grid/src/WorklistDataGrid.jsx
@@ -212,7 +212,9 @@ function WorklistDataGrid(props) {
   const [focusedCol, setFocusedCol] = useState(0);
   const [ariaLiveMessage, setAriaLiveMessage] = useState(null);
   const [cellAriaLiveMessage, setCellAriaLiveMessage] = useState(null);
-  const rowSelection = useRef({ isRowSelectionModeToggledByGrid: false });
+  const isRowSelectionModeToggledByGrid = useRef(false);
+  const inShiftUpDownMode = useRef(false);
+  const multiSelectRange = useRef({ start: null, end: null });
 
   // Define ColumnContext Provider value object
   const columnContextValue = useMemo(() => ({ pinnedColumnOffsets, setCellAriaLiveMessage }), [pinnedColumnOffsets]);
@@ -279,15 +281,14 @@ function WorklistDataGrid(props) {
     setFocusedRowCol(newFocusCell.row, newFocusCell.col, false);
 
     if (!hasSelectableRows) {
-      rowSelection.current.rangeUsingFocusedRow = null;
-      rowSelection.current.rangeUsingLastIndividuallySelectedRow = null;
+      multiSelectRange.current = {};
     }
 
     // Since the row selection mode has changed, the row selection mode needs to be updated.
-    if (!rowSelection.current.isRowSelectionModeToggledByGrid) {
+    if (!isRowSelectionModeToggledByGrid.current) {
       setAriaLiveMessage(intl.formatMessage({ id: hasSelectableRows ? 'Terra.worklist-data-grid.row-selection-mode-enabled' : 'Terra.worklist-data-grid.row-selection-mode-disabled' }));
     }
-    rowSelection.current.isRowSelectionModeToggledByGrid = false;
+    isRowSelectionModeToggledByGrid.current = false;
 
     setDataGridColumns(displayedColumns.map((column) => initializeColumn(column)));
   // eslint-disable-next-line react-hooks/exhaustive-deps
@@ -335,7 +336,7 @@ function WorklistDataGrid(props) {
   );
 
   const handleClearRowSelection = () => {
-    rowSelection.current.rangeUsingLastIndividuallySelectedRow = null;
+    multiSelectRange.current = {}; // Clear the information used for selecting multiple rows.
     if (isRowSelected()) {
       setAriaLiveMessage(intl.formatMessage({ id: 'Terra.worklist-data-grid.all-rows-unselected' }));
       // Esc (while in row selection mode and rows are selected): Clear selection
@@ -343,7 +344,6 @@ function WorklistDataGrid(props) {
         onClearSelectedRows();
       }
     } else if (onDisableSelectableRows) {
-      rowSelection.current.rangeUsingFocusedRow = null;
       setAriaLiveMessage(intl.formatMessage({ id: 'Terra.worklist-data-grid.row-selection-mode-disabled' }));
       onDisableSelectableRows();
     }
@@ -357,37 +357,46 @@ function WorklistDataGrid(props) {
     }
   };
 
-  const selectMultipleRows = useCallback((newEndRowIndex, rangeInfo) => {
-    if (!onRowSelect || !rangeInfo) {
-      return;
+  const getRowsInRangeToUnselect = useCallback((startOfRange, prevEndOfRange, newEndOfRange) => {
+    let rowIdsToUnselect = [];
+    if (!prevEndOfRange) {
+      // short-circuit since there was no previous range, there is nothing to unselect.
+      return rowIdsToUnselect;
     }
 
+    if (startOfRange > prevEndOfRange) {
+      // The range extends upward from the anchor row
+      if (newEndOfRange > prevEndOfRange) {
+        // The range was moved down towards the anchor so rows that no longer qualify for the range need to be unselected.
+        rowIdsToUnselect = rows.slice(prevEndOfRange - 1, Math.min(newEndOfRange, startOfRange)).map(row => row.id);
+      }
+    } else if (startOfRange < prevEndOfRange) {
+      // The range extends downward from the anchor row
+      if (newEndOfRange < prevEndOfRange) {
+        // The range was moved up towards the anchor so rows that no longer qualify for the range need to be unselected. New endRangeIndex becomes ordered End
+        rowIdsToUnselect = rows.slice(Math.max(startOfRange, newEndOfRange), prevEndOfRange).map(row => row.id);
+      }
+    }
+    return rowIdsToUnselect;
+  }, [rows]);
+
+  const selectMultipleRows = useCallback((startOfRange, prevEndOfRange, newEndOfRange) => {
     if (!hasSelectableRows) {
       onEnableRowSelection();
     }
+    if (!onRowSelect) {
+      return;
+    }
 
     // Rows in range to remain selected
-    const selectionStartRowIndex = Math.min(rangeInfo.anchorRow, newEndRowIndex);
-    const selectionEndRowIndex = Math.max(rangeInfo.anchorRow, newEndRowIndex);
+    const selectionStartRowIndex = Math.min(startOfRange, newEndOfRange);
+    const selectionEndRowIndex = Math.max(startOfRange, newEndOfRange);
 
     // We are subtracting 1 to accommodate for the column header in the grid.
     const rowIdsToSelect = rows.slice(selectionStartRowIndex - 1, selectionEndRowIndex).map(row => row.id);
 
     // Determine if there are rows that are no longer in range that need to be unselected.
-    let rowIdsToUnselect = [];
-    if (rangeInfo.anchorRow > rangeInfo.previousSelectionEndRow) {
-      // The range extends upward from the anchor row
-      if (newEndRowIndex > rangeInfo.previousSelectionEndRow) {
-        // The range was moved down towards the anchor so rows that no longer qualify for the range need to be unselected.
-        rowIdsToUnselect = rows.slice(rangeInfo.previousSelectionEndRow - 1, Math.min(newEndRowIndex, rangeInfo.anchorRow)).map(row => row.id);
-      }
-    } else if (rangeInfo.anchorRow < rangeInfo.previousSelectionEndRow) {
-      // The range extends downward from the anchor row
-      if (newEndRowIndex < rangeInfo.previousSelectionEndRow) {
-        // The range was moved up towards the anchor so rows that no longer qualify for the range need to be unselected. New endRangeIndex becomes ordered End
-        rowIdsToUnselect = rows.slice(Math.max(rangeInfo.anchorRow, newEndRowIndex), rangeInfo.previousSelectionEndRow).map(row => row.id);
-      }
-    }
+    const rowIdsToUnselect = getRowsInRangeToUnselect(startOfRange, prevEndOfRange, newEndOfRange);
 
     let ariaMessage = intl.formatMessage({
       id: hasSelectableRows
@@ -399,22 +408,26 @@ function WorklistDataGrid(props) {
       id: rowIdsToSelect.length === 1
         ? 'Terra.worklist-data-grid.row-selection-template'
         : 'Terra.worklist-data-grid.row-selection-selected-rows-range',
-    }, { row: rangeInfo.anchorRow, endRow: newEndRowIndex })}`;
+    }, { row: startOfRange, endRow: newEndOfRange })}`;
 
     setAriaLiveMessage(ariaMessage);
     onRowSelect(rowIdsToSelect, rowIdsToUnselect);
-  }, [hasSelectableRows, intl, onEnableRowSelection, onRowSelect, rows]);
+  }, [hasSelectableRows, intl, onEnableRowSelection, onRowSelect, rows, getRowsInRangeToUnselect]);
 
   const selectRow = useCallback((rowId, rowIndex) => {
     const rowsToSelect = [];
     const rowsToUnSelect = [];
-    rowSelection.current.rangeUsingLastIndividuallySelectedRow = { anchorRow: rowIndex };
     const rowLabel = rows[rowIndex - 1].ariaLabel || (rowIndex + 1);
 
     if (!rows[rowIndex - 1].isSelected) {
+      multiSelectRange.current = { start: rowIndex, end: null }; // Establish new starting point for future range.
       setAriaLiveMessage(intl.formatMessage({ id: 'Terra.worklist-data-grid.row-selection-template' }, { row: rowLabel }));
       rowsToSelect.push(rowId);
     } else {
+      if (rowIndex === multiSelectRange.current?.start) {
+        // The row that denotes the start for multiselect has been cleared.
+        multiSelectRange.current = {};
+      }
       setAriaLiveMessage(intl.formatMessage({ id: 'Terra.worklist-data-grid.row-selection-cleared-template' }, { row: rowLabel }));
       rowsToUnSelect.push(rowId);
     }
@@ -459,33 +472,21 @@ function WorklistDataGrid(props) {
     setFocusedRow(selectionDetails.rowIndex);
     setFocusedCol(selectionDetails.columnIndex);
 
-    const rangeInfo = selectionDetails.selectedByKeyboard ? rowSelection.current.rangeUsingLastIndividuallySelectedRow : rowSelection?.current?.rangeUsingFocusedRow;
-    if (!selectionDetails.multiSelect) {
-      if (hasSelectableRows) {
-        // regular click or space key
-        selectRow(selectionDetails.rowId, selectionDetails.rowIndex);
+    if (!hasSelectableRows) {
+      if (selectionDetails.isShiftKeyPressed) {
+        // Shift+Space, Shift+Click, Shift+Up, Shift+Down
+        isRowSelectionModeToggledByGrid.current = !hasSelectableRows;
+        multiSelectRange.current = { start: selectionDetails.rowIndex, end: null };
+        selectMultipleRows(multiSelectRange.current.start, multiSelectRange.current.end, selectionDetails.rowIndex);
+        multiSelectRange.current.end = selectionDetails.rowIndex;
       }
-    } else if (hasSelectableRows) {
-      // Shift key is pressed
-      if (rangeInfo?.anchorRow) {
-        // ACTION: select multiple rows based on anchor
-        selectMultipleRows(selectionDetails.rowIndex, rangeInfo);
-        rangeInfo.previousSelectionEndRow = selectionDetails.rowIndex;
-      } else {
-        // no anchor row
-        selectRow(selectionDetails.rowId, selectionDetails.rowIndex);
-      }
-    } else if (selectionDetails.selectedByKeyboard) {
-      // Shift + Space
-      rowSelection.current.isRowSelectionModeToggledByGrid = !hasSelectableRows;
-      rowSelection.current.rangeUsingLastIndividuallySelectedRow = { anchorRow: selectionDetails.rowIndex };
-      selectMultipleRows(selectionDetails.rowIndex, rowSelection.current.rangeUsingLastIndividuallySelectedRow);
-      rowSelection.current.rangeUsingLastIndividuallySelectedRow.previousSelectionEndRow = selectionDetails.rowIndex;
+    } else if (selectionDetails.isShiftKeyPressed && multiSelectRange.current?.start) {
+      // Select multiple rows based on previously established anchor.
+      selectMultipleRows(multiSelectRange.current.start, multiSelectRange.current.end, selectionDetails.rowIndex);
+      multiSelectRange.current.end = selectionDetails.rowIndex;
     } else {
-      // Shift + Click
-      rowSelection.current.isRowSelectionModeToggledByGrid = !hasSelectableRows;
-      selectMultipleRows(selectionDetails.rowIndex, rangeInfo);
-      rangeInfo.previousSelectionEndRow = selectionDetails.rowIndex;
+      // There is no anchor or the shift key is not pressed so select the row and establish an anchor.
+      selectRow(selectionDetails.rowId, selectionDetails.rowIndex);
     }
   }, [hasSelectableRows, selectMultipleRows, selectRow]);
 
@@ -493,7 +494,7 @@ function WorklistDataGrid(props) {
     const key = event.keyCode;
     switch (key) {
       case KeyCode.KEY_SHIFT:
-        rowSelection.current.rangeUsingFocusedRow = null;
+        inShiftUpDownMode.current = false;
         break;
       default:
         break;
@@ -548,11 +549,6 @@ function WorklistDataGrid(props) {
           activateMultiRowSelection = true;
         }
         nextRow += 1;
-        break;
-      case KeyCode.KEY_SHIFT:
-        if (cellCoordinates.row !== 0 && !rowSelection.current?.rangeUsingFocusedRow?.anchorRow) {
-          rowSelection.current.rangeUsingFocusedRow = { anchorRow: cellCoordinates.row, previousSelectionEndRow: cellCoordinates.row };
-        }
         break;
       case KeyCode.KEY_LEFT:
         if (event.metaKey) {
@@ -637,9 +633,13 @@ function WorklistDataGrid(props) {
       } else if (nextRow <= 0) {
         nextRow = 1; // Only non-header rows can be selected.
       }
-      rowSelection.current.isRowSelectionModeToggledByGrid = !hasSelectableRows;
-      selectMultipleRows(nextRow, rowSelection.current.rangeUsingFocusedRow);
-      rowSelection.current.rangeUsingFocusedRow.previousSelectionEndRow = nextRow;
+      if (!inShiftUpDownMode.current) {
+        inShiftUpDownMode.current = true;
+        multiSelectRange.current = { start: cellCoordinates.row, end: null };
+      }
+      isRowSelectionModeToggledByGrid.current = !hasSelectableRows;
+      selectMultipleRows(multiSelectRange.current.start, multiSelectRange.current.end, nextRow);
+      multiSelectRange.current.end = nextRow;
     }
     // Handle the normal case
     if (nextRow > rows.length || nextCol >= displayedColumns.length) {