+++ packages/terra-worklist-data-grid/tests/jest/WorklistDataGrid.test.jsx
@@ -40,6 +40,15 @@ const dataFile = {
         { content: '2.28', isMasked: true },
       ],
     },
+    {
+      id: '4',
+      isSelected: true,
+      cells: [
+        { content: 'Oxygen Flow Rate (L/min)' },
+        { content: '63' },
+        { content: '47' },
+      ],
+    },
   ],
 };
 
@@ -238,10 +247,12 @@ describe('basic grid', () => {
 
 describe('with pinned columns', () => {
   it('sets pinnedColumns as pinned', () => {
+    const pinnedColumns = dataFile.cols.slice(0, 2);
+
     const wrapper = mountWithIntl(
       <WorklistDataGrid
         id="test-terra-worklist-data-grid"
-        pinnedColumns={dataFile.cols.slice(0, 2)}
+        pinnedColumns={pinnedColumns}
         overflowColumns={dataFile.cols.slice(2)}
         rows={dataFile.rows}
       />,
@@ -249,14 +260,16 @@ describe('with pinned columns', () => {
 
     const pinnedColumnHeaderCells = wrapper.find('.pinned');
 
-    expect(pinnedColumnHeaderCells).toHaveLength(2 * 4);
+    expect(pinnedColumnHeaderCells).toHaveLength(pinnedColumns.length * (dataFile.rows.length + 1));
   });
 
   it('sets row selection column as pinned', () => {
+    const pinnedColumns = dataFile.cols.slice(0, 2);
+
     const wrapper = mountWithIntl(
       <WorklistDataGrid
         id="test-terra-worklist-data-grid"
-        pinnedColumns={dataFile.cols.slice(0, 2)}
+        pinnedColumns={pinnedColumns}
         overflowColumns={dataFile.cols.slice(2)}
         rows={dataFile.rows}
         hasSelectableRows
@@ -265,7 +278,7 @@ describe('with pinned columns', () => {
 
     const pinnedColumnHeaderCells = wrapper.find('.pinned');
 
-    expect(pinnedColumnHeaderCells).toHaveLength(3 * 4);
+    expect(pinnedColumnHeaderCells).toHaveLength((pinnedColumns.length + 1) * (dataFile.rows.length + 1));
   });
 
   it('pins row selection column if pinnedColumns is undefined', () => {
@@ -280,7 +293,7 @@ describe('with pinned columns', () => {
 
     const pinnedColumnHeaderCells = wrapper.find('.pinned');
 
-    expect(pinnedColumnHeaderCells).toHaveLength(1 * 4);
+    expect(pinnedColumnHeaderCells).toHaveLength(1 * (dataFile.rows.length + 1));
     expect(console.warn).toHaveBeenCalledWith(expect.stringContaining(ERRORS.PINNED_COLUMNS_UNDEFINED)); // eslint-disable-line no-console
   });
 });
@@ -322,5 +335,475 @@ describe('Error handling - prop types', () => {
     ).dive();
 
     expect(console.error).toHaveBeenCalledWith(expect.stringContaining(ERRORS.ROW_HEADER_INDEX_EXCEEDS_PINNED)); // eslint-disable-line no-console
+  });
+});
+
+describe('Row selection', () => {
+  it('verifies only onRowSelect is called when space is used to select a row.', () => {
+    const mockOnCellSelect = jest.fn();
+    const mockOnRowSelect = jest.fn();
+    const mockOnEnableRowSelection = jest.fn();
+
+    const wrapper = mountWithIntl(
+      <WorklistDataGrid
+        id="test-terra-worklist-data-grid"
+        pinnedColumns={dataFile.cols.slice(0, 2)}
+        overflowColumns={dataFile.cols.slice(2)}
+        rows={dataFile.rows}
+        hasSelectableRows
+        onRowSelect={mockOnRowSelect}
+        onEnableRowSelection={mockOnEnableRowSelection}
+        onCellSelect={mockOnCellSelect}
+      />,
+    );
+
+    // Find and select a selectable cell in the 3rd row.
+    const selectableCell = wrapper.find(Row).at(2).find('td.selectable');
+    selectableCell.at(0).simulate('keydown', { keyCode: 32 });
+
+    // Validate that only the onRowSelect callback is called.
+    expect(mockOnRowSelect).toHaveBeenCalledWith(['3'], []);
+    expect(mockOnEnableRowSelection).not.toHaveBeenCalled();
+    expect(mockOnCellSelect).not.toHaveBeenCalled();
+
+    expect(wrapper).toMatchSnapshot();
+  });
+
+  it('verifies only onRowSelect is called when mouse is used to select a row.', () => {
+    const mockOnCellSelect = jest.fn();
+    const mockOnRowSelect = jest.fn();
+    const mockOnEnableRowSelection = jest.fn();
+
+    const wrapper = mountWithIntl(
+      <WorklistDataGrid
+        id="test-terra-worklist-data-grid"
+        pinnedColumns={dataFile.cols.slice(0, 2)}
+        overflowColumns={dataFile.cols.slice(2)}
+        rows={dataFile.rows}
+        hasSelectableRows
+        onRowSelect={mockOnRowSelect}
+        onEnableRowSelection={mockOnEnableRowSelection}
+        onCellSelect={mockOnCellSelect}
+      />,
+    );
+
+    // Find and select a selectable cell in the 3rd row.
+    const selectableCell = wrapper.find(Row).at(2).find('td.selectable');
+    selectableCell.at(0).simulate('mouseDown');
+
+    // Validate that only the onRowSelect callback is called.
+    expect(mockOnRowSelect).toHaveBeenCalledWith(['3'], []);
+    expect(mockOnEnableRowSelection).not.toHaveBeenCalled();
+    expect(mockOnCellSelect).not.toHaveBeenCalled();
+
+    expect(wrapper).toMatchSnapshot();
+  });
+
+  it('verifies only onRowSelect is called when space is used to unselect a row.', () => {
+    const mockOnCellSelect = jest.fn();
+    const mockOnRowSelect = jest.fn();
+    const mockOnEnableRowSelection = jest.fn();
+
+    const wrapper = mountWithIntl(
+      <WorklistDataGrid
+        id="test-terra-worklist-data-grid"
+        pinnedColumns={dataFile.cols.slice(0, 2)}
+        overflowColumns={dataFile.cols.slice(2)}
+        rows={dataFile.rows}
+        hasSelectableRows
+        onRowSelect={mockOnRowSelect}
+        onEnableRowSelection={mockOnEnableRowSelection}
+        onCellSelect={mockOnCellSelect}
+      />,
+    );
+
+    // The 4th row is selected so unselect it using space key.
+    const selectableCell = wrapper.find(Row).at(3).find('td.selectable');
+    selectableCell.at(0).simulate('keydown', { keyCode: 32 });
+
+    // Validate that only the onRowSelect callback is called.
+    expect(mockOnRowSelect).toHaveBeenCalledWith([], ['4']);
+    expect(mockOnEnableRowSelection).not.toHaveBeenCalled();
+    expect(mockOnCellSelect).not.toHaveBeenCalled();
+
+    expect(wrapper).toMatchSnapshot();
+  });
+
+  it('verifies only onRowSelect is called when mouse is used to unselect a row.', () => {
+    const mockOnCellSelect = jest.fn();
+    const mockOnRowSelect = jest.fn();
+    const mockOnEnableRowSelection = jest.fn();
+
+    const wrapper = mountWithIntl(
+      <WorklistDataGrid
+        id="test-terra-worklist-data-grid"
+        pinnedColumns={dataFile.cols.slice(0, 2)}
+        overflowColumns={dataFile.cols.slice(2)}
+        rows={dataFile.rows}
+        hasSelectableRows
+        onRowSelect={mockOnRowSelect}
+        onEnableRowSelection={mockOnEnableRowSelection}
+        onCellSelect={mockOnCellSelect}
+      />,
+    );
+
+    // The 4th row is selected so unselect it using mouse.
+    const selectableCell = wrapper.find(Row).at(3).find('td.selectable');
+    selectableCell.at(0).simulate('mouseDown');
+
+    // Validate that only the onRowSelect callback is called.
+    expect(mockOnRowSelect).toHaveBeenCalledWith([], ['4']);
+    expect(mockOnEnableRowSelection).not.toHaveBeenCalled();
+    expect(mockOnCellSelect).not.toHaveBeenCalled();
+
+    expect(wrapper).toMatchSnapshot();
+  });
+
+  it('verifies callbacks when Shift+Down is used and row selection mode is not enabled.', () => {
+    const mockOnCellSelect = jest.fn();
+    const mockOnRowSelect = jest.fn();
+    const mockOnEnableRowSelection = jest.fn();
+
+    const wrapper = mountWithIntl(
+      <WorklistDataGrid
+        id="test-terra-worklist-data-grid"
+        pinnedColumns={dataFile.cols.slice(0, 2)}
+        overflowColumns={dataFile.cols.slice(2)}
+        rows={dataFile.rows}
+        onRowSelect={mockOnRowSelect}
+        onEnableRowSelection={mockOnEnableRowSelection}
+        onCellSelect={mockOnCellSelect}
+      />,
+    );
+
+    const selectableCell = wrapper.find(Row).at(0).find('td.selectable');
+    selectableCell.at(0).simulate('mouseDown'); // Row selection is not on so cell will be selected.
+    expect(mockOnCellSelect).toHaveBeenCalledWith('1', 'Column-1'); // The first click to select the cell from which shift+Down will occur.
+
+    selectableCell.at(0).simulate('keydown', { shiftKey: true, keyCode: 40 });
+    expect(mockOnRowSelect).toHaveBeenCalledWith(['1', '2'], []);
+    expect(mockOnEnableRowSelection).toHaveBeenCalled();
+
+    expect(wrapper).toMatchSnapshot();
+  });
+
+  it('verifies callbacks when Shift+Down is used and row selection mode is enabled.', () => {
+    const mockOnCellSelect = jest.fn();
+    const mockOnRowSelect = jest.fn();
+    const mockOnEnableRowSelection = jest.fn();
+
+    const wrapper = mountWithIntl(
+      <WorklistDataGrid
+        id="test-terra-worklist-data-grid"
+        pinnedColumns={dataFile.cols.slice(0, 2)}
+        overflowColumns={dataFile.cols.slice(2)}
+        rows={dataFile.rows}
+        hasSelectableRows
+        onRowSelect={mockOnRowSelect}
+        onEnableRowSelection={mockOnEnableRowSelection}
+        onCellSelect={mockOnCellSelect}
+      />,
+    );
+
+    const selectableCell = wrapper.find(Row).at(0).find('td.selectable');
+    selectableCell.at(0).simulate('keydown', { shiftKey: true, keyCode: 40 });
+
+    expect(mockOnRowSelect).toHaveBeenCalledWith(['1', '2'], []);
+    expect(mockOnEnableRowSelection).not.toHaveBeenCalled();
+    expect(mockOnCellSelect).not.toHaveBeenCalled();
+
+    expect(wrapper).toMatchSnapshot();
+  });
+
+  it('verifies callbacks when Shift+Down more than one.', () => {
+    const mockOnCellSelect = jest.fn();
+    const mockOnRowSelect = jest.fn();
+    const mockOnEnableRowSelection = jest.fn();
+
+    const wrapper = mountWithIntl(
+      <WorklistDataGrid
+        id="test-terra-worklist-data-grid"
+        pinnedColumns={dataFile.cols.slice(0, 2)}
+        overflowColumns={dataFile.cols.slice(2)}
+        rows={dataFile.rows}
+        hasSelectableRows
+        onRowSelect={mockOnRowSelect}
+        onEnableRowSelection={mockOnEnableRowSelection}
+        onCellSelect={mockOnCellSelect}
+      />,
+    );
+
+    const selectableCell = wrapper.find(Row).at(0).find('td.selectable');
+    selectableCell.at(0).simulate('keydown', { shiftKey: true, keyCode: 40 });
+    selectableCell.at(0).simulate('keydown', { shiftKey: true, keyCode: 40 });
+
+    expect(mockOnRowSelect).toHaveBeenCalledWith(['1', '2'], []);
+    expect(mockOnRowSelect).toHaveBeenCalledWith(['1', '2', '3'], []);
+    expect(mockOnEnableRowSelection).not.toHaveBeenCalled();
+    expect(mockOnCellSelect).not.toHaveBeenCalled();
+
+    expect(wrapper).toMatchSnapshot();
+  });
+
+  it('verifies multiple independent ranges can be created with Shift+Down.', () => {
+    const mockOnCellSelect = jest.fn();
+    const mockOnRowSelect = jest.fn();
+    const mockOnEnableRowSelection = jest.fn();
+
+    const wrapper = mountWithIntl(
+      <WorklistDataGrid
+        id="test-terra-worklist-data-grid"
+        pinnedColumns={dataFile.cols.slice(0, 2)}
+        overflowColumns={dataFile.cols.slice(2)}
+        rows={dataFile.rows}
+        hasSelectableRows
+        onRowSelect={mockOnRowSelect}
+        onEnableRowSelection={mockOnEnableRowSelection}
+        onCellSelect={mockOnCellSelect}
+      />,
+    );
+
+    // Find a cell on row 1 and use Shift+Down to create a selected range of rows.
+    let selectableCell = wrapper.find(Row).at(0).find('td.selectable');
+    selectableCell.at(0).simulate('keydown', { shiftKey: true, keyCode: 40 });
+    expect(mockOnRowSelect).toHaveBeenCalledWith(['1', '2'], []);
+
+    // Release the Shift key and Arrow Down once to get to row 3.
+    selectableCell.at(0).simulate('keyup', { keyCode: 16 });
+    selectableCell.at(0).simulate('keydown', { shiftKey: false, keyCode: 40 });
+
+    // Find a cell on Row 3 and do a Shift+Down
+    selectableCell = wrapper.find(Row).at(2).find('td.selectable');
+    selectableCell.at(0).simulate('keydown', { shiftKey: true, keyCode: 40 });
+    expect(mockOnRowSelect).toHaveBeenCalledWith(['3', '4'], []);
+
+    expect(mockOnEnableRowSelection).not.toHaveBeenCalled();
+    expect(mockOnCellSelect).not.toHaveBeenCalled();
+
+    expect(wrapper).toMatchSnapshot();
+  });
+
+  it('verifies callbacks when Shift+Down contracts the range of selected rows.', () => {
+    const mockOnCellSelect = jest.fn();
+    const mockOnRowSelect = jest.fn();
+    const mockOnEnableRowSelection = jest.fn();
+
+    const wrapper = mountWithIntl(
+      <WorklistDataGrid
+        id="test-terra-worklist-data-grid"
+        pinnedColumns={dataFile.cols.slice(0, 2)}
+        overflowColumns={dataFile.cols.slice(2)}
+        rows={dataFile.rows}
+        hasSelectableRows
+        onRowSelect={mockOnRowSelect}
+        onEnableRowSelection={mockOnEnableRowSelection}
+        onCellSelect={mockOnCellSelect}
+      />,
+    );
+
+    const selectableCell = wrapper.find(Row).at(0).find('td.selectable');
+    selectableCell.at(0).simulate('keydown', { shiftKey: true, keyCode: 40 }); // Shift+Down starts the range
+    selectableCell.at(0).simulate('keydown', { shiftKey: true, keyCode: 40 }); // Shift+Down extends the range
+    selectableCell.at(0).simulate('keydown', { shiftKey: true, keyCode: 38 }); // Shift+Up contracts the range
+
+    expect(mockOnRowSelect).toHaveBeenCalledWith(['1', '2'], []); // range is established with 2 rows selected.
+    expect(mockOnRowSelect).toHaveBeenCalledWith(['1', '2', '3'], []); // range is expanded with an additional row selected.
+    expect(mockOnRowSelect).toHaveBeenCalledWith(['1', '2'], ['3']); // range is contracted with with one row unselected.
+    expect(mockOnEnableRowSelection).not.toHaveBeenCalled();
+    expect(mockOnCellSelect).not.toHaveBeenCalled();
+
+    expect(wrapper).toMatchSnapshot();
+  });
+
+  it('verifies Shift+Click honors the anchor established by row selection using space bar.', () => {
+    const mockOnCellSelect = jest.fn();
+    const mockOnRowSelect = jest.fn();
+    const mockOnEnableRowSelection = jest.fn();
+
+    const wrapper = mountWithIntl(
+      <WorklistDataGrid
+        id="test-terra-worklist-data-grid"
+        pinnedColumns={dataFile.cols.slice(0, 2)}
+        overflowColumns={dataFile.cols.slice(2)}
+        rows={dataFile.rows}
+        hasSelectableRows
+        onRowSelect={mockOnRowSelect}
+        onEnableRowSelection={mockOnEnableRowSelection}
+        onCellSelect={mockOnCellSelect}
+      />,
+    );
+
+    let selectableCell = wrapper.find(Row).at(0).find('td.selectable');
+    // Select Row 1 using space.
+    selectableCell.at(0).simulate('keydown', { keyCode: 32 });
+    expect(mockOnRowSelect).toHaveBeenCalledWith(['1'], []);
+
+    // Shift+Click on row 3
+    selectableCell = wrapper.find(Row).at(2).find('td.selectable');
+    selectableCell.at(0).simulate('mouseDown', { shiftKey: true });
+    expect(mockOnRowSelect).toHaveBeenCalledWith(['1', '2', '3'], []);
+
+    expect(mockOnEnableRowSelection).not.toHaveBeenCalled();
+    expect(mockOnCellSelect).not.toHaveBeenCalled();
+
+    expect(wrapper).toMatchSnapshot();
+  });
+
+  it('verifies Shift+Space honors the anchor established by row selection using Mouse.', () => {
+    const mockOnCellSelect = jest.fn();
+    const mockOnRowSelect = jest.fn();
+    const mockOnEnableRowSelection = jest.fn();
+
+    const wrapper = mountWithIntl(
+      <WorklistDataGrid
+        id="test-terra-worklist-data-grid"
+        pinnedColumns={dataFile.cols.slice(0, 2)}
+        overflowColumns={dataFile.cols.slice(2)}
+        rows={dataFile.rows}
+        hasSelectableRows
+        onRowSelect={mockOnRowSelect}
+        onEnableRowSelection={mockOnEnableRowSelection}
+        onCellSelect={mockOnCellSelect}
+      />,
+    );
+
+    let selectableCell = wrapper.find(Row).at(0).find('td.selectable');
+    // Select Row 1 using Mouse.
+    selectableCell.at(0).simulate('mouseDown');
+    expect(mockOnRowSelect).toHaveBeenCalledWith(['1'], []);
+
+    // Shift+Space on row 3
+    selectableCell = wrapper.find(Row).at(2).find('td.selectable');
+    selectableCell.at(0).simulate('keyDown', { shiftKey: true, keyCode: 32 });
+    expect(mockOnRowSelect).toHaveBeenCalledWith(['1', '2', '3'], []);
+
+    expect(mockOnEnableRowSelection).not.toHaveBeenCalled();
+    expect(mockOnCellSelect).not.toHaveBeenCalled();
+
+    expect(wrapper).toMatchSnapshot();
+  });
+
+  it('verifies Shift+Click enables row selection when row selection is turned off.', () => {
+    const mockOnCellSelect = jest.fn();
+    const mockOnRowSelect = jest.fn();
+    const mockOnEnableRowSelection = jest.fn();
+
+    const wrapper = mountWithIntl(
+      <WorklistDataGrid
+        id="test-terra-worklist-data-grid"
+        pinnedColumns={dataFile.cols.slice(0, 2)}
+        overflowColumns={dataFile.cols.slice(2)}
+        rows={dataFile.rows}
+        onRowSelect={mockOnRowSelect}
+        onEnableRowSelection={mockOnEnableRowSelection}
+        onCellSelect={mockOnCellSelect}
+      />,
+    );
+
+    const selectableCell = wrapper.find(Row).at(2).find('td.selectable');
+    // Select Row 3 using Shift+Click when row selection is off.
+    selectableCell.at(0).simulate('mouseDown', { shiftKey: true });
+    expect(mockOnRowSelect).toHaveBeenCalledWith(['3'], []);
+    expect(mockOnEnableRowSelection).toHaveBeenCalled();
+
+    expect(mockOnCellSelect).not.toHaveBeenCalled();
+
+    expect(wrapper).toMatchSnapshot();
+  });
+
+  it('verifies Shift+Space enables row selection when row selection is turned off.', () => {
+    const mockOnCellSelect = jest.fn();
+    const mockOnRowSelect = jest.fn();
+    const mockOnEnableRowSelection = jest.fn();
+
+    const wrapper = mountWithIntl(
+      <WorklistDataGrid
+        id="test-terra-worklist-data-grid"
+        pinnedColumns={dataFile.cols.slice(0, 2)}
+        overflowColumns={dataFile.cols.slice(2)}
+        rows={dataFile.rows}
+        onRowSelect={mockOnRowSelect}
+        onEnableRowSelection={mockOnEnableRowSelection}
+        onCellSelect={mockOnCellSelect}
+      />,
+    );
+
+    const selectableCell = wrapper.find(Row).at(2).find('td.selectable');
+    // Select Row 3 using Shift+Space when row selection is off.
+    selectableCell.at(0).simulate('keyDown', { shiftKey: true, keyCode: 32 });
+    expect(mockOnRowSelect).toHaveBeenCalledWith(['3'], []);
+    expect(mockOnEnableRowSelection).toHaveBeenCalled();
+
+    expect(mockOnCellSelect).not.toHaveBeenCalled();
+
+    expect(wrapper).toMatchSnapshot();
+  });
+
+  it('verifies Shift+Click establishes anchor and selects row when row selection is already enabled.', () => {
+    const mockOnCellSelect = jest.fn();
+    const mockOnRowSelect = jest.fn();
+    const mockOnEnableRowSelection = jest.fn();
+
+    const wrapper = mountWithIntl(
+      <WorklistDataGrid
+        id="test-terra-worklist-data-grid"
+        pinnedColumns={dataFile.cols.slice(0, 2)}
+        overflowColumns={dataFile.cols.slice(2)}
+        rows={dataFile.rows}
+        hasSelectableRows
+        onRowSelect={mockOnRowSelect}
+        onEnableRowSelection={mockOnEnableRowSelection}
+        onCellSelect={mockOnCellSelect}
+      />,
+    );
+
+    let selectableCell = wrapper.find(Row).at(0).find('td.selectable');
+    // Select Row 1 using Shift+Click. Row is selected and anchor created.
+    selectableCell.at(0).simulate('mouseDown', { shiftKey: true });
+    expect(mockOnRowSelect).toHaveBeenCalledWith(['1'], []);
+
+    // Shift+Click on row 3
+    selectableCell = wrapper.find(Row).at(2).find('td.selectable');
+    selectableCell.at(0).simulate('mouseDown', { shiftKey: true });
+    expect(mockOnRowSelect).toHaveBeenCalledWith(['1', '2', '3'], []);
+
+    expect(mockOnEnableRowSelection).not.toHaveBeenCalled();
+    expect(mockOnCellSelect).not.toHaveBeenCalled();
+
+    expect(wrapper).toMatchSnapshot();
+  });
+
+  it('verifies Shift+Space establishes anchor and selects row if row selection is already enabled.', () => {
+    const mockOnCellSelect = jest.fn();
+    const mockOnRowSelect = jest.fn();
+    const mockOnEnableRowSelection = jest.fn();
+
+    const wrapper = mountWithIntl(
+      <WorklistDataGrid
+        id="test-terra-worklist-data-grid"
+        pinnedColumns={dataFile.cols.slice(0, 2)}
+        overflowColumns={dataFile.cols.slice(2)}
+        rows={dataFile.rows}
+        hasSelectableRows
+        onRowSelect={mockOnRowSelect}
+        onEnableRowSelection={mockOnEnableRowSelection}
+        onCellSelect={mockOnCellSelect}
+      />,
+    );
+
+    let selectableCell = wrapper.find(Row).at(0).find('td.selectable');
+    // Select Row 1 using Shift+Space. Row is selected and anchor created.
+    selectableCell.at(0).simulate('keyDown', { shiftKey: true, keyCode: 32 });
+    expect(mockOnRowSelect).toHaveBeenCalledWith(['1'], []);
+
+    // Shift+Space on row 3
+    selectableCell = wrapper.find(Row).at(2).find('td.selectable');
+    selectableCell.at(0).simulate('keyDown', { shiftKey: true, keyCode: 32 });
+    expect(mockOnRowSelect).toHaveBeenCalledWith(['1', '2', '3'], []);
+
+    expect(mockOnEnableRowSelection).not.toHaveBeenCalled();
+    expect(mockOnCellSelect).not.toHaveBeenCalled();
+
+    expect(wrapper).toMatchSnapshot();
   });
 });